UX UI DESING PROJECT DESCRIPTION
This project is based on the creation of a wedding app of an existing wedding brand. The brand is Rue de Seine. Their principal problem is that they only sell online. Customers can try their dresses in specific ateliers that are located in big cities. I discovered the opportunity to create an app where customers from small cities that can not travel could be able to buy a wedding dress. A very important, personal and expensive purchase.